+++ packages/cli-testing-library/src/index.ts
@@ -25,13 +25,13 @@ if (
 }
 
 export * from './config'
-export * from './helpers.js'
-export * from './events.js'
-export * from './get-queries-for-instance.js'
-export * from './matches.js'
+export * from './helpers'
+export * from './events'
+export * from './get-queries-for-instance'
+export * from './matches'
 export * from './pure'
 export * from './query-helpers'
-export * from './queries/index.js'
-export * from './mutation-observer.js'
-export * from './wait-for.js'
+export * from './queries/index'
+export * from './mutation-observer'
+export * from './wait-for'
 export * from './user-event'
