--- packages/cli-testing-library/types/index.d.ts
@@ -1,17 +1,0 @@
-// TypeScript Version: 3.8
-
-import * as queries from './queries'
-import * as queryHelpers from './query-helpers'
-
-export {queries, queryHelpers}
-
-export * from './config'
-export * from './events'
-export * from './get-queries-for-instance'
-export * from './matches'
-export * from './pure'
-export * from './queries'
-export * from './query-helpers'
-export * from './suggestions'
-export * from './wait-for'
-export * from './user-event/index'
